+++ drivers/media/usb/uvc/uvc_ctrl.c
@@ -24,6 +24,7 @@
 #include <linux/workqueue.h>
 #include <linux/atomic.h>
 #include <media/v4l2-ctrls.h>
+#include <media/v4l2-uvc.h>
 
 #include "uvcvideo.h"
